+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - `f-breadcrumb` 组件 `separator` 配置项改为 `VNode` 类型，默认值修改为 `null`
 - `f-notification` 组件 `close-btn` 参数默认值改为 `null`
 - `f-icon` 组件永久废除
+- `f-button` 组件 `before-icon` `loading-icon` `after-icon` 配置项改为 `VNode` 类型，默认值修改为 `null`
 
 ## 0.7.0-alpha.9 (2022-10-19)